v1
View Pointclouds in VR, use the oculus remote to move around scene and monipulate the cloud from a GUI (radius, rotation).
Problems: there is a slight lag when messing with rotation and person is close to the cloud (this is because of the way the orientation is maintained. Need to find a better way)